doc/requirements.txt
aiida-core==2.5.1
aio-pika==6.8.1
aiormq==3.3.1
alabaster==0.7.16
alembic==1.13.1
annotated-types==0.6.0
anyascii==0.3.2
appdirs==1.4.4
archive-path==0.4.2
ase==3.22.1
astroid==3.0.3
asttokens==2.4.1
async-generator==1.10
attrs==23.2.0
Babel==2.14.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bleach==6.1.0
boto3==1.34.50
botocore==1.34.50
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
circus==0.18.0
click==8.1.7
click-spinner==0.1.10
comm==0.2.1
contourpy==1.2.0
cryptography==42.0.4
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
disk_objectstore==1.0.0
docstring-parser==0.15
docutils==0.20.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
fonttools==4.49.0
future==1.0.0
graphviz==0.20.1
greenlet==3.0.3
h5py==3.10.0
idna==3.6
imagesize==1.4.1
importlib-metadata==6.11.0
ipykernel==6.29.2
ipython==8.22.1
ipywidgets==8.1.2
jedi==0.18.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.10
kiwipy==0.7.8
kiwisolver==1.4.5
latexcodec==2.0.1
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mistune==3.0.2
monty==2024.2.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nbsphinx==0.9.3
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
packaging==23.2
palettable==3.3.3
pamqp==2.3.0
pandas==2.2.0
pandocfilters==1.5.1
paramiko==2.12.0
parso==0.8.3
pexpect==4.9.0
pgsu==0.2.4
phonopy==2.21.2
pillow==10.2.0
platformdirs==4.2.0
plotly==5.19.0
plumpy==0.21.10
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pybtex==0.24.0
pycparser==2.21
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic_core==2.16.2
Pygments==2.17.2
pymatgen==2024.2.23
PyNaCl==1.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytray==0.3.4
pytz==2021.3
pyxtal==0.6.3
PyYAML==6.0.1
pyzmq==25.1.2
qmpy-rester==0.2.0
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
scikit-learn==1.4.1.post1
scipy==1.12.0
shortuuid==1.0.11
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.5
spglib==2.3.1
Sphinx==7.2.6
sphinx-autoapi==3.0.0
sphinx-immaterial==0.11.10
sphinx_design==0.5.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-contentui==0.2.5
sphinxcontrib-details-directive==0.1.0
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.27
stack-data==0.6.3
sympy==1.12
tabulate==0.8.10
tenacity==8.2.3
threadpoolctl==3.3.0
tinycss2==1.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2024.1
uncertainties==3.1.7
upf-to-json==0.9.5
urllib3==2.0.7
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.10
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
-e .
